[ID-128] Completing verification for countries api given filter

+++ src/test/java/api/model/Filter.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Filter {
-    private double gdp;
+    private float gdp;
     private String operator;
 }
